Move TemplateProcessor to a separate module
