+++ packages/advisor-components/src/TemplateProcessor/index.ts
@@ -1,0 +1,2 @@
+export * from './TemplateProcessor';
+export { default as TemplateProcessor } from './TemplateProcessor';
